+++ internal/lax509/README.md
@@ -1,4 +1,4 @@
-# x509fork
+# lax509
 
 This is a minimalist fork of [`crypto/x509`](https://pkg.go.dev/crypto/x509).
