+++ internal/lax509/x509.go
@@ -1,4 +1,4 @@
-package x509fork
+package lax509
 
 var (
 	oidExtensionSubjectAltName = []int{2, 5, 29, 17}
